DynamicInputStream now creates ArraySources on demand.
So far the append() method only supported ArraySource objects. Now it is
possible to pass a plain array which is wrapped automatically in an
ArraySource. There is also a convenience method for creating a source for an
array. A conversion from an array to an array source is frequently needed;
therefore, it makes sense to have this available at the DynamicInputStream
class.