Introduced a new MediumID trait.
The purpose is to simplify the handling of media and their files and to stop
the separation between normal media and a list of other files. A
MediaScanResult now only consists of a single map; the additional list for
other files is no longer necessary.

Note that this commit breaks the build because other classes have to be
adapted to this change.