Reworked DirectoryScanner to fix a problem when assigning files to media.
The original algorithm was relying on the fact that files are visited before
subdirectories. This is not the case; therefore, another algorithm had to be
written. Now all media and escription files are retrieved in a first step. The
assignment of media files to meda is done in a second step.