LocalBufferActor now always expects a buffer manager to be passed in.
It does not really makes sense to create a default buffer manager because the
location of the buffer directory is use case specific.

The current behavior caused sproadic test failures: The buffer directory was
set to the temp directory, and tests that tried to clean this directory could
fail if there were currently open temporary files.