Added DirectoryScanner class.
This is a more generic version of the DirectoryScanner class from the media
package. The functionality for scanning directory structures is needed by other
components, too. Therefore, it has been extracted in a form that can be used in
a broader context. The media.DirectoryScanner class is going to be reworked in
terms of the new class.