Added new action tasks for appending songs to the playlist.
So far, these actions have only be used by the MediumPopupHandler class.
In future, they will be added to the menu and toolbar. Therefore, the
implementation of action tasks has to be reworked: the tasks are not set
dynamically, but they have to dynamically obtain the songs to be added.