Added a threshold for loading new data into the audio buffer.
The audio buffer was always filled when it contained less data than
its capacity. Now a fill operation is only performed when a full
audio chunk fits in.

This is an optimization. Normally, because the audio stream is
wrapped by a buffered stream, the number of load operations is
significantly reduced. However, it happend that another load
operation was triggered immediately when the buffer could not be
fully filled.