Added some more integration-related functionality to SourceStreamReaderActor.
The actor can now notify a listener about the actual audio stream to
be played by sending an AudioSource message. This is needed on player
side to have a correct stream URL (e.g. when creating the playback
context).

A delegating supervisor strategy has been set, so that every exception
in a child actor can be handled by the parent.