Updated configurations of actor system.
The update to a newer version of akka also brought some changes in
the configuration of akka-remote. As this file is currently copied,
the copy had to be updated.

The regular configuration was extended by the declaration of a
special dispatcher for actors doing blocking stuff. This is needed
for the audio player engine.