Enhancements of ClientApplication.
The class now stores the ClientApplicationContext as bean in its
context. It publishes a special initialized message when its
initialization is done.

The purpose of these means is to have a dedicated life-cycle for
listeners registered on the message bus: A listener registered by an
application can now find out when the application is up and running.