Added classes for loading resources for the configuration script.
When executing the Jelly script for the media interface configuration
the resources cannot be resolved because the standard resource
manager does not take the changed class loader into account.

This could be fixed by setting an alternative resource manager, but
unfortunately, the resource manager classes from JGUIraffe are not
available in an OSGi environment because they are located under an
impl package. So as a temporary fix some new resource classes have
been introduced. But this is only a hack; they will be removed when
the next version of JGUIraffe becomes available.