ApplicationStartup now supports configuration naming.
The startup method now expects an application name. From this name
the name of the main configuration and the user configuration are
derived. There are some mechanisms to adapt the latter.

The goal of this functionality is to support multiple installations
of LineDJ applications that can either share their configurations or
use separate ones.