ClientApplication now triggers the consumer registration.
Analogous to the automatic registration of message bus listeners, a
bean of type ConsumerRegistrationProcessor is checked to exist in the
main window's bean context. If it is found, it is initialized which
causes the consumer registration to happen automatically.

The application no longer explicitly requests the current state of
the media archive connection. This is now handled via the
ArchiveAvailableExtension.