Added AvailableMediaExtension class.
This is an extension of the media archive interface which manages an
AvailableMedia object. It allows platform components to keep track
easily on the media available in the archive.

Note that the current implementation is not 100 percent complete.
There will be some changes in the registration of state listeners.
This area has to be reworked as there are currently some race
conditions.