Reworked state listener registration in RelayActor.
The actor manages now a set of component IDs for the registered
listeners. It is responsible for adding and removing a registration
with the media archive based on the number of listeners available.
When the connection to the archive is lost and established anew,
a listener registration is created if necessary.

Note that the extension components created so far have to be adapted
to this new behavior.