Extended meta data requests by a registration ID.
There can be race conditions which clients that add and remove meta
data listeners: A client may register a meta data listener and later
remove the registration. After a while another listener for the same
medium is registered. On the archive, data for this medium may become
available and a chunk (containing only the incremental part of data)
is sent to the client shortly before the first listener is removed.
This data may be received by the 2nd listener registration, and the
listener has no chance to detect that it is incremental for another
listener.

To avoid this, meta data listener registrations now contain a numeric
listener registration ID. This ID is also included in meta data
response messages. If clients correctly increment registration IDs,
they can detect stale response messages.

So far only the registration ID has been added to message classes.
It is not yet evaluated or correctly initialized.