Reworked MediaFileUriHandler.
The class is now only an object with utility methods (it is not
really necessary to create instances). URIs for the undefined
medium now also contain the archive component ID. It is now
possible to extract the single components of such an URI.

This functionality is needed by the meta data union actor to fully
support the global undefined medium list.