Introduced sequence numbers in HTTP archive messages.
These are used to cancel scan operations in a graceful way: The
HTTP archive management actor creates a unique sequence number for
each scan operation. When an operation is canceled or one is
completed the sequence number is increased. If then further
results arrive with an out-dated sequence number, they can be
ignored.