Enhancements of TempFileActorManager.
Changed the return value of downloadCompletedArrived() to return more
information about the current read operation. This is required by the
owning download actor to do corresponding clean-up.

For this purpose, some new internal data classes have been introduced.

There is also a new method that returns a list of temporary files that
have been written, but not yet read.