Added a helper trait for executing single HTTP requests.
This trait uses a Flow as created by HttpFlowFactory to execute single
requests. It returns a Future with the response that fails if a
processing error occurs or the response returned from the server was not
successful.

This trait is going to be used by multiple actors that need to send HTTP
requests.