Introducing a state for HTTP archives.
In the past it was not easily possible for a client to determine the
state an HTTP archive is in. For instance, if the request for the
archive's content file failed, this was not exposed to the outside.

Now a hierarchy of HttpArchiveState classes has been introduced. The
management actor for an HTTP archive supports new message that allow
querying the current state.