ClientApplication now calls Application.shutdown on deactivation.
It has to be ensured that all cleanup is correctly executed, even if
the application is closed in the hard way (e.g. by stopping the
bundle). Therefore, ClientApplication now invokes the shutdown()
method explicitly in its deactivation method. JGUIraffe handles the
case that this method has already been called before.