Fixed some problems with HttpArchiveStartupApplication.
Registration of bus listeners is done differently to avoid a potential
NPE when a notification arrives before the config manager has been
setup.

The correct request message is now sent to the archive management
actor.

When new credentials for a realm arrive the state of this realm is
reset, and all associated archives are logged out if necessary. This
is required because otherwise the changed credentials would not be
applied.