Changed access to system properties in configuration files.
When multiple LineDJ applications are contained in a deployment there
were from time to time errors related to the creation of configuration
files. These errors indicated a concurrent modification while reading
a map configuration. Root cause was adding new system properties (for
the name of the application's configurations) during application
startup.

To avoid this, system properties in configuration files are now
accessed via the sys: variable prefix. This should not cause a new
MapConfiguration to be created, but a direct access to the map with
system properties.