Introduced notion of an activated playlist.
It is now possible to add songs to the playlist, but not activating
them; i.e. the songs are not yet passed to the player engine. This is
useful if the playlist may change before playback actually starts, e.
g. when songs are only added to a playlist editor.

As soon as songs are passed to the player engine, steps are taken to
copy or download song files. If the playlist is not yet finalized,
such steps should not be done.