Added a module to manage the state of the scan sink.
The stream to scan a media archive processes the content of media and
their description files in parallel. As soon as both information for
one or multiple media is available, it should be passed downstream,
respecting back-pressure.

As the logic to manage this state is complex, it has been extracted
into a module the provides corresponding state update functions.