Added properties to the media archive configuration.
There are a few new properties related to scanning of media files. The
documentation has been updated as well.

The representation of roots in MediaArchiveConfig has been reworked
as well: In future only a single directory to be parsed is supported.
Therefore, code could be simplified, and no hierarchical
configuration is needed any more.