Some enhancements of MediaScanStateUpdateService.
The service now updates MediumInfo objects to have the correct
checksum (which is calculated by another component). There are new
functions to handle a scan cancellation or a scan completed event. In
both cases, it may be necessary to send messages to other actors;
therefore, an object with transition messages is made available here.