Fixed a problem with time calculation for sources partly skipped.
If the playback time for a chunk of data could be calculated from the
audio format, the playback time displayed by the player was doubled.
This is because it was initialized from the skip time of the source
and then updated while reaching the skip position. This bug has been
fixed.