Added ProcessingResultCombiningStage.
As tests have the shown, partial processing results for an HTTP
archive can arrive out of order from the HTTP flow stage. Therefore,
it is not sufficient to simply zip them. This new flow stage is now
responsible for combining partial results to final results even when
their order is incorrect. The results that belong together are matched
by their medium IDs.