MediumInfoParser now allows specifying a checksum.
The parser used to set the checksum of the resulting MediumInfo object
to an empty string. In some use cases, the checksum is already known
when the parser is invoked. Therefore, the value to be used for the
checksum can now be passed as an argument when invoking the parser.