MediaFileUriHandler now generates relative path URIs.
URIs used by media archives should be local and not specific to
concrete implementations. Therefore, the proprietary 'path://' prefix
has been removed.

A new function has been added to find the medium ID for the specific
undefined medium a song belongs to. This is going to be needed by the
browser application to make sure that songs on the global undefined
medium are handled correctly.

MediaFileUriHandler now also performs URL encoding.