Added two more MetaDataStateEvent types.
The new events are used to wrap multiple meta data scan operations.
After the archive components have been rewritten to add results to the
archive as soon as they are available, they trigger multiple scan
events. The new events are an option to ignore such partial results
and get only notified when the whole update operation is done.