Changed type definition of ValidationFlow.
The flow gets now a list with the files of a complete medium as input.
It can then decide whether to validate the single files or split them
up to other groups like albums or the medium as a whole. This is
easier to handle than grouping a sequence of files together.