Added StreamPullModeratorActor.
This is a base class for actors that manage a StreamPullReadService to
enable request-based access to a push-based stream. The actor sets
itself as the sink of a stream and coordinates data from the stream
with requests of its clients. Concrete sub classes have to define the
message protocol to be used in the communication with the clients.