Made PlayerFacadeActor more generic regarding playback sources.
The actor now supports a SourceActorCreator function. This function is
invoked when creating the source reader actor needed by the managed
playback actor. The actors returned by this function are also managed
by PlayerFacadeActor; they are closed, stopped, and re-created as
necessary.

Background is that this actor is going to be used by the radio player
as well. Then a different source reader actor needs to be managed.