RadioSchedulerActor no longer changes the source currently played.
The actor now only monitors the current radio source and sends out
events if a replacement source should be played. It is up to the
receiver of the event to actually switch the source.

This was done to decouple the scheduler from the radio source actor;
so it can focus on its main task: changing to a replacement source if
necessary. Another reason is that changing the radio source to be
played is going to become more complicated in future.

Two event types have been introduced related to radio replacement
sources.