Added ShutdownManagementActor class.
This actor implementation is going to play an important role in the
managed shutdown process in future. It monitors the registered
shutdown observers and waits for their confirmations.

The reason why this can no longer be done by the ShutdownHandler is
that the UI thread in this phase is obviously no longer active.
Therefore, another mechanism to synchronize shutdown receivers need to
be found. The advantage of the actor is that it is relatively easy to
implement a shutdown timeout.