Reworked ShutdownHandler to use a ShutdownManagementActor.
Rather than managing incoming shutdown confirmations on the event bus
itself, the handler now delegates monitoring of shutdown observers to
a ShutdownManagementActor instance.

As the message bus may no longer be functional at this point of time,
interaction with shutdown observers is now done directly via method
calls or sending messages to the shutdown management actor. (This
changed the interfaces of some of the message classes involved.)