Added a new crypt project handling encryption and decryption.
Copied the CryptStage class from the stream-sync project and its test
class.

So far, only decryption is supported by LineDJ. In future, to
implement the "super password" feature for HTTP archives, encryption
is needed, too. This functionality is provided by the new classes
copied over. The goal is to place the classes supporting encryption
and decryption in a dedicated project from where they can be accessed
easily.