Added a function to format long durations to DurationTransformer.
This is going to be used by the archive admin application to display
the total playback duration of the songs contained in the archive.
Here, a hour-based representation is typically of limited value. So,
for long durations, a string with the number of days, hours, minutes,
and seconds is generated.