Added a trait to download media files.
This trait is going to replace the HttpArchiveProtocol trait in future
by making use of implementations based on the CloudFiles project. The
only relevant interaction of an HTTP archive with a protocol is
downloading of media files. The new trait handles exactly this.

Adapted build.sbt to add CloudFiles dependencies to the HTTP
archive project; also the OSGi exports have been adjusted.