Added support for typed actors to ActorManagement trait.
Regarding actor management, typed actors are a bit special: They
cannot be stopped in a generic way, and querying their correctly typed
reference is hard. ActorManagement has now been extended to support
the registration of typed actors by introducing an ActorStopper trait.
This trait has to be implemented in a way specific to a typed actor.
The trait then makes sure that this stopper is invoked when the
component is deactivated.