Added HttpArchiveStartupConfig class.
This class is the new representation of the configuration for an HTTP
archive required to startup the archive.

The existing HttpArchiveConfig class is going to be stripped down to
the properties that are needed for the management of the archive.
Properties only required for starting up the archive have been moved
to the new startup configuration class.