Reworked HttpArchiveStarter to use the new factories.
The starter now creates a fully initialized MediaDownloader and adds
it to the archive configuration that is passed to the actors created
for the archive. A new startup() function has been added that
implements the reworked startup procedure.

In the following, HttpStartupApplication needs to be reworked to call
this new function.