Removed obsolete fields from HttpArchiveConfig.
The fields that are also contained in HttpArchiveStartupConfig are no
longer needed. The class becomes a plain data class, the functionality
to extract an instance from a Configuration is obsolete as well.

This causes some updates on client classes that were setting these
fields, typically in tests. There were also some wrong references to
property constants pointing to HttpArchiveConfig.