Removed RequestActorTestImpl test helper class.
This class was used to test interactions with a request actor, which
is no longer needed, since HTTP requests are done via the CloudFiles
library.

The test helper class used to provide a function to create a test
configuration for an archive. This code has been inlined into the spec
classes that were using it.