Switched to Scala 2.13.
The JGUIraffe dependency is now available for this version of Scala;
therefore, the update is possible. This caused some further updates in
dependencies to fix conflicts SBT was complaining about, especially
in regard to scala-modules.

Adaptations on the source code were required to make the project
compile again. These were mostly related to the conversion of Java
collections to Scala.

Currently, the project compiles, but there are still test failures
caused by conflicting versions of Akka.