Reworked dependencies of mp3PlaybackContextFactory module.
The module now declares the dependencies to Aries as well. The SpiFly
classifier is disabled, so that the processed artifact is used
directly. The soundlibs have been upgraded to their most recent
versions.

In OSGi images, this project can now be referenced as a dependency,
and then all the required dependencies to play mp3 audio are
automatically pulled in. The additional base OSGi image
"baseAudioPlatform" is no longer necessary.