Dealt with Apache Commons Logging.
Redirecting the output produced by Commons Logging (which is used by
JGUIraffe) to Log4j is, unfortunately, not trivial. The approach with
a fragment bundle did not work, because Commons Logging was not able
to find its configuration file in this fragment bundle. The file could
only be found if the Thread-context class loader was explicitly set.

Therefore, this logic (and the custom configuration file setting a
LogFactory specific to Log4j) has been implemented in the platform
module, in ClientManagementApplication, which is the first application
class that gets instantiated.

This works for all client application classes started later; only the
logs of ClientManagementApplication itself are not propagated to
Log4j.