Tweaked the OsgiImagePlugin again.
There seemed to be a race condition for bundles processed by SpiFly:
sometimes the processed bundle and sometimes the original bundle was
added to the OSGi image. The order in which artifacts for tasks are
queried has been changed, in the hope that now the SpiFly task is
executed first - which includes the osgiBundle task. So this task
should not be executed again.