HttpArchiveContentProcessorActor: Changed component IDs.
In MediumID instances created for HTTP archives the archive's URI is
used as component ID. These URIs, however, are not necessarily unique:
There are cases where the same URIs are used and only user credentials
make the difference. Requests to download media files can fail if the
hosting archive cannot be identified.

To make component IDs unique, the archive name is now used.