MediaMessages: Added more data to MediumFiles message.
The set with URIs is going to be replaced by a set with MediaFileID
objects containing more information about media files. That way, it is
easier to request specific media files from the archive, without
having to do complex URI mappings.

The set with URIs is going to be removed from the message once all
producers and consumers have been reworked.