MetaDataUnionActor: Improved metadata requests during scans.
When receiving a request for metadata while a scan is in progress and
the medium is not known, an UnknownMedium response is only sent back
if no listener is registered. Otherwise, the listener is registered;
if data for this medium arrives later, the listener is notified.

It is also no longer possible to register a listener for the undefined
medium if no scan is in progress.