MetaDataCache: Added the nested MediumContent class.
This class is going to be used as a representation of a medium and its
content instead of MetaDataChunk. It uses MediaFileIDs as identifiers
for files and can therefore be used to store the synthetic global
undefined medium.

A follow-up commit will use this class as type for the MetaDataCache
extension.