Archive: Introduced a MediumChecksum value class.
Checksums for media were used frequently in the archive project. As
they were represented as plain strings, it was not obvious in the
code, which values were actually checksums. This change significantly
increases the readability in these cases.

Note that this class is used only internally in the archive module; in
data objects accessed by other modules, checksums are still
represented as strings.