PersistentMetaDataWriterActor: A URI mapping is no longer used.
When writing metadata files the actor no longer relies on a URI-file
mapping to obtain the paths of files in progress. It now uses a
PathUriConverter for this purpose.

The ProcessMedium message has been updated to contain such a converter
rather than a mapping. The senders of this message still have to be
adapted.