Archive: Some actors now expect a PathUriConverter to be passed in.
The main actors to manage a local archive now expect a
PathUriConverter to be passed to the constructor. This converter is
going to be used instead of a path-URI-mapping in order to obtain
paths for URIs and vice versa.

ArchiveActorFactory has been adapted accordingly. It now creates a
correct converter and passes it to the actors.