MediaManagerActor: Changed resolving of media files.
Media files are no longer looked up in the scan state, but directly in
the local folder structure the requested URI points to. The goal is to
get rid of the mapping constructing during the scan. This is possible,
since the URI of a media file can be directly converted to its local
path.