MediaScanStateUpdateService: Changed file data in scan state.
Information about the files contained in the archive is now stored as
a map containing the URIs of the media files for each medium. This
data can be updated easily for incoming results without having to deal
with URI mappings. To compute URIs for media files, a function
parameter is used; here the URI conversion function of a
PathUriConverter is passed in.

MediaManagerActor has been adapted to this change.