MediaScannerActor: Changed the creation of MediumIDs.
The URI and the path to the settings file of a MediumID are now
relative to the root of the archive. That way, they contain no longer
local paths of specific archives and are thus more generic.

The IDs are created by MediumAggregateStage and MediaScannerActor;
both have been adapted. The creation function is now located in
MediumAggregateStage.