ArchiveToCWriterActor: Simplified writing of archive ToCs.
The actor no longer applies an ArchiveContentTableConfig to do a
complex URI mapping. Rather, URIs to media description files are
written directly, and for metadata files only the names are stored. It
is possible to correctly resolve these files when the root paths for
the media and metadata files of an archive are known.