Http archives: Extended configuration.
HttpArchiveConfig now defines additional properties for the paths to
the content file, the media files, and the metadata files. The archive
URI now only defines the base path, the property therefore has been
renamed.

With this change, it is easier to resolve the relative URIs that are
now written in the archive ToC files and in metadata files.