MediaManagerActor: Improved handling of queries for media.
In the MediumIDs passed in requests, the archiveComponentID is no
longer evaluated when matching against the data stored in the archive.
This makes it possible to download data from the archive that was
hosted before by a different archive.

To achieve this, media contained in the archive are stored with a
MediumID without an archiveComponentID. For the MediumIDs in incoming
requests, the property is dropped. The MediaScanStateUpdateService
still has to be updated to produce the correct scan state.