TimeoutAwareHttpDownloadActor: Removed MinimumTimeoutDelay logic.
The race condition that a timeout message is received after the
schedule was canceled can be prevented by checking whether the
cancelable property is defined.

The approach used previously (based on a constant for the minimum
delay between messages) was problematic when the timeout is configured
close or even below this constant. Then timeouts were ignored.