Added StoppableTestProbe helper class.
This class can be used as a replacement for Akka's TestProbe class
when collaborator actors need to be stopped in tests (e.g. to test
that death watch is applied correctly). It has been discovered that
stopping normal test probes can have undesired effects on the
stability of the test actor system. Therefore, this helper class is
introduced for testing such scenarios.