TempFileActorManager: Reworked the creation of TempReadOperations.
The manager is no longer responsible for creating and managing reader
actors and TempReadOperation objects. This functionality is now
abstracted by a new trait TempReadOperationHolder. This simplifies the
code, and it is not necessary to deal with a ChildActorFactory.

The change is also motivated by the fact that
TimeoutAwareHttpDownloadActor is going to need access to the current
reader actor. Therefore, it makes sense, that this actor class creates
the child reader actor itself and then provides it to the
TempFileActorManager helper object.