TimeoutAwareHttpDownloadActor: Changed tests for sender actors.
When receiving DownloadDataResult or DownloadComplete messages it has
to be checked whether these messages stem from a reader actor for
temporary files or the wrapped download actor. The conditions used
before no longer work, since the download actor is not the source of
such messages (internally, another wrapped actor is used). Therefore,
the conditions now check for the reader actor, which is known after
implementing the TempReadOperationHolder trait.