M3uReader: Resolving is now only done if necessary.
The class now checks the file extension of the passed in stream
reference. Only if it refers to a playlist file, an attempt to resolve
the audio stream URI is made. Otherwise, the reference is returned
without changes. With this change, the whole logic to resolve audio
stream URIs is now handled by this class.

Signed-off-by: oheger <oliver.heger@oliver-heger.de>