build: Updated SbtSpiFly plugin.
The new plugin version 0.4.0 provides a skipSpiFly setting. This
allows using this plugin for all OSGi bundles. The build has been
changed accordingly. (All projects except for
mp3PlaybackContextFactory set this new flag to true.)

The background of this change is that the OsgiImagePlugin now only
have to query for SbtSpiFly tasks. This hopefully resolves race
conditions leading to invalid SpiFly bundles.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>