OsgiImagePlugin: Further updates to make it running reliably.
Running only the SpiFly task caused corruption on affected bundles. To
fix this, now the osgiBundle and publishLocal tasks are run on
dependent projects. Then the published artifacts are retrieved, and
the correct files are copied into the OSGi image.

In the settings of the SpiFly plugin, now the default classifier is
used, so that the artifact produced by this plugin can be identified.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>