Player engine: Introduced a new RadioEvent hierarchy.
The player events have been split into events raised by the audio
player and events of the radio player. This is done as a preparation
to extract the radio player engine from the player engine.

The resulting solution (and event handling in general) is not ideal
though, since the processing is not type-safe. This needs to be
revisited later.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>