PlayerControl: Added a function to create event manager actors.
This is a first step to convert the event management to the new
EventManagerActor implementation. Until clients are reworked to the
type-safe event model, both an old and a new event manager actor are
available. The new function allows creating both. In addition, it
connects the actors, so that events published via the new event
manager are also received from old event listeners.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>