RadioController: The player is now passed via the message bus.
This replaces the mechanism of storing the player in the bean context.
In the future, the initialization of the radio player will be done
asynchronously; therefore, the player cannot be stored by the
application in the bean context when it is started. It will become
available at a later point of time and must then be propagated to the
controller.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>