Audio platform: Added new AudioPlayerManagerActor class.
This actor is going to be responsible for the asynchronous creation of
an audio player instance. With the current implementation, this is
rather complex, as PlaybackContextFactories being added or removed in
a separate thread have to be taken into account. Moving this logic
into an actor implementation simplifies the handling of concurrency.

In a follow-up commit, the actor will be used by
AudioPlatformComponent.

To avoid cyclic dependencies between packages, AudioPlayerController
has been moved into the actors package.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>