AudioPlatformComponent: Fixed a race condition wrt metadata resolving.
The metadata resolver service needs to be registered with the audio
controller to receive events when there are state changes. Due to the
asynchronous creation of the controller, the registration message was
already published on the message bus before the controller was created
and set up. This has been fixed by letting AudioPlayerManagerActor
handle the publishing of messages to the controller.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>