Radio player engine: Added RadioEventConverterActor class.
This is an actor implementation that can be registered as an event
listener for audio player events and converts these events to
corresponding radio events. It is going to be needed when the
player engines have been reworked to generate type-safe events. Then
some events generated by the playback actor must be converted to radio
events.

The hierarchy of radio events has been extended to mirror some missing
events of the audio player.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>