Player engine: Added a test ActorCreator implementation.
This implementation already handles stub actors related to the event
manager actor. It can be extended by further checks for other actors.
Dealing with stub actors simulating the event management actor is
tricky, and it is needed for both the audio player and the radio
player. Therefore, it makes sense to have the logic at a central
place. It has been extracted from RadioPlayerSpec.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>