Radio player engine: Introduced an actor to manage the radio player.
RadioPlayerManagerActor is analogous to AudioPlayerManagerActor from
the audio platform. It creates the radio player asynchronously and
manages playback context factories that can arrive or go away at any
time.

PlayerManagerActor was extended by a callback for a failed player
creation, since in this case, RadioPlayerManagerActor needs to send a
corresponding message on the event bus.

The current solution is a bit hacky though, because the creation of
the player has to be synchronized with the initialization of the
player UI. This may change in the future: The radio player could be
accessed via a commands sent via the message bus in the same way as
this is done for the audio player.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>