RadioPlayerApplication: Again changed configuration access.
In 38baf60c, the main configuration was passed to the controller
instead of the user configuration. This caused the issue that the
current radio source was no longer saved.

Now the user configuration is passed again, but the main configuration
(which contains the definition of the radio sources) is accessed via
the application context.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>