RadioController: Added support for displaying metadata.
The UI has a new text field to display metadata about the current
song. The controller offers a new method to be invoked when receiving
a metadata event.

This is just a quick solution to display metadata in the UI. Some
tweaks are likely necessary in the future, e.g. to deal with longer
texts.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>