Platform: Added TextTimeFunctions module.
This object provides functionality to generate texts that depend on
elapsed time. This is going to be used in the UI of platform
applications, e.g. to show playback time or generate rotating text for
song titles.

The functionality was already available partly in the
CurrentSongController class. It has been extracted and generalized, so
that it can be used in a broader scope.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>