PlayerConfig: Added a threshold for time progress events.
Currently, playback progress events are generated every second. The
purpose of this new property is to make this configurable, so that
multiple events per second (or less) can be generated. The meaning is
that another event can be fired when the accumulated playback time
after the last event exceeds this threshold.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>