ScanResultEnhancer: Fixed a problem with checksum generation.
The algorithm did not handle the relative medium root paths correctly.
This caused the path to be resolved against the current directory; so
different checksums have been generated depending on the folder in
which the archive was launched. Now relative paths are always resolved
against the root of the scan, and the checksum is then calculated on
the correct relative file URIs.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>