MetaDataManagerActor: Fixed a race condition.
It could happen that metadata files were not written because the
PersistentMetaDataWriterActor registered at the metadata union actor
before the media contribution message for the affected medium arrived.
The registration was then rejected because the medium was unknown.

The problem could be fixed by changing the order in which messages
were sent.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>