RadioPlayer: Changed representation of exclusion queries.
The interval queries defining exclusions for radio sources are now
longer passed as a map, but as a function. This is a more flexible
approach.

With the previous mechanism, there was also the danger that radio
sources that did not define exclusion queries were dropped, since they
were not part of the map.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>