Radio player engine: Added RadioSourceStateService trait.
This service manages the state of the currently played radio source.
It checks whether the source reaches an exclusion interval and can
then switch to a replacement source. Later on it is checked whether
playback can switch again to the original source.

The current implementation is not yet complete; further functionality
is going to be added later.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>