Player engine: Added ScheduledInvocationActor.
This actor implementation provides very basic scheduler functionality
by sending a message to a target actor after a given delay. This can
be used to implement all kinds of (single-shot) scheduled invocations.
It is also going to be used by DelayActor, which is more specific to
use cases of the player engine.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>